move views module into directory to enable splitting
file was getting a bit unwieldy so start process of splitting it up
